+++ src/behaviors/cameraAvStreamManagementServer.ts
@@ -68,6 +68,7 @@ export class MatterbridgeCameraAvStreamManagementServer extends CameraAvStreamMa
   CameraAvStreamManagement.Feature.Video,
   CameraAvStreamManagement.Feature.Audio,
   CameraAvStreamManagement.Feature.Snapshot,
+  CameraAvStreamManagement.Feature.Speaker,
   CameraAvStreamManagement.Feature.ImageControl,
 ) {
   /**